Created five module files
