+++ index.js
@@ -3,105 +3,11 @@
 const collection = document.querySelector('#book-collection');
 
 
-// main class
-class Main {
-  static displayBooks() {
-    // eslint-disable-next-line no-use-before-define
-    const books = Store.getBooks();
-
-    books.forEach((book) => Main.addBookToList(book));
-  }
-
-  static addBookToList(book) {
-    const element = document.createElement('div');
-    element.classList.add('item');
-    element.setAttribute('id', book.title);
-
-    element.innerHTML = `
-    <p id="author">"${book.title}" by ${book.author}</p>
-    <button type="button" class="delete">Remove</button>`;
-
-    collection.appendChild(element);
-  }
-
-  static deleteBook(el) {
-    if (el.classList.contains('delete')) {
-      el.parentElement.remove();
-    }
-  }
-
-  static clearFields() {
-    document.querySelector('#new-title').value = '';
-    document.querySelector('#new-author').value = '';
-  }
-}
-
-// Store Class
-class Store {
-  static getBooks() {
-    let books;
-    if (localStorage.getItem('books') === null) {
-      books = [];
-    } else {
-      books = JSON.parse(localStorage.getItem('books'));
-    }
-    return books;
-  }
-
-  static addBook(book) {
-    const books = Store.getBooks();
-
-    books.push(book);
-
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-
-  static removeBook(index) {
-    const books = Store.getBooks();
-    let counter = 0;
-
-    books.forEach((book) => {
-      if (book.title === index) {
-        books.splice(counter, 1);
-      }
-      counter += 1;
-    });
-
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-}
 
 // Event: Display Books
-// Event: Add a book
-document.querySelector('#add').addEventListener('click', (e) => {
-  // prevent actual submit
-  e.preventDefault();
 
-  // Get form values
-  const title = document.querySelector('#new-title').value;
-  const author = document.querySelector('#new-author').value;
 
-  // instatiate book
-  const book = new Book(title, author);
 
-  // Add Book to Main
-  Main.addBookToList(book);
-
-  // Add book to store
-  Store.addBook(book);
-
-  // Clear field
-  Main.clearFields();
-});
-
-// Event: Remove a book
-collection.addEventListener('click', (e) => {
-// From Main
-  Main.deleteBook(e.target);
-
-  // From Store
-  Store.removeBook(e.target.parentElement.id);
-});
 
 // Full Website
 // List
@@ -116,26 +22,7 @@ const formSection = document.querySelector('.form');
 const item3 = document.querySelector('#nav-item3');
 const contactSection = document.querySelector('#contact');
 
-// list function
-item1.addEventListener('click', () => {
-  bookSection.style.display = 'block';
-  formSection.style.display = 'none';
-  contactSection.style.display = 'none';
-});
 
-// Add new function
-item2.addEventListener('click', () => {
-  formSection.style.display = 'block';
-  bookSection.style.display = 'none';
-  contactSection.style.display = 'none';
-});
-
-// Contact function
-item3.addEventListener('click', () => {
-  contactSection.style.display = 'block';
-  bookSection.style.display = 'none';
-  formSection.style.display = 'none';
-});
 
 const dayTime = document.querySelector('#date');
 dayTime.innerText = new Date();
